Removed logout
There is no need for a logout function because there is no simple way to destroy the token. Instead, the frontend should simply discard the token in order to logout.

I also wrote tests for expired tokens, which I found out was not actually working. I fixed it and it now works.